feat: RIR girişi tıklayınca açılan renkli kaydırıcı paneli (#266)
- RIR artık yazılmaz; alana dokununca on duraklı kaydırıcı açılır:
  0, 0–1, 1 … 4, 4+. Ray accent'ten sönüğe degrade, altında durağın
  başlığı ve cümlesi, ⓘ ile RIR açıklaması, eklemede Temizle.
- Web ve mobil aynı özellik (RirAlani); ortak hesap
  packages/shared/src/lib/rir.ts, metinler tr/en katalogda.
- Backend: SetEntry.Rir int -> numeric(4,1), yarım adımlı (2.5 = 2–3),
  0–5 aralığı ve RirScale ile yarım adım kuralı; export "RIR 2–3" / "4+".
- Eski 5 üstü RIR silinmez, "4+" gösterilir; düzenleyici onu 5 olarak açar.
- Düzenleyicide Temizle yok (PATCH RIR'i boşaltamaz).

Refs #266

+++ src/Grind.Api/Models/Entities/SetEntry.cs
@@ -16,8 +16,11 @@ public class SetEntry
     /// <summary>Bu set kaydedildiği anda bir rekor kırdı mı — tarihsel anlık görüntü.</summary>
     public RecordType RecordType { get; set; }
 
-    /// <summary>Reps in Reserve.</summary>
-    public int? Rir { get; set; }
+    /// <summary>
+    /// Reps in Reserve — yarım adımlı (#266): 2.5 = "2–3 arası", 5 = "4+". #266 öncesi kayıtlarda
+    /// 5'ten büyük olabilir; gösterimde "4+" sayılır.
+    /// </summary>
+    public decimal? Rir { get; set; }
 
     public DateTime CreatedAt { get; set; }
